Implement champion system for battles
A champion will be picked on each side for each battle and they will
duel. Champions are chosen by rolling the new knight stat "bravery".
The loser dies and the winner earns an advantage for their side
in battle. Loads of glory is awarded to the winners of duels.

Also refactored some stuff into cleaner functions.