Split project up
Attempted to clean the project up a bit by moving the bulk of it
into a "game" module. Within the moduel logical groupings of
code are split out into their own files. Its not perfect and
won't be for a while but it should make it easier to work on
in  the time being.