Implement wars
Implemented the first pass on wars. Houses with a
high enough tension will declare war and other houses
will join them as allies if they have a common hatred
of the target. The war will run in a series of battles
over the seasons until one side's morale is reduced to 0.
Winners are rewards with +1 might and losers are punished
with -1 might. Plenty of TODOs were left along the way for
future improvemnts.