Implement first pass on marriage
If houses are not at war with each other the player
is able to marry a knight from one house into another.
The knight from the weaker house will marry into the
stronger house.

Will likely need to add a glory cost to marriage, and
consider adding a wealth transfer as compensation to the
weaker house.